Update to session-checks - return value => echo
The return was replaced with echo as the variable assignment doesn't work when a value is returned and is supposed to be assigned to it. So the old
if ! ${dur:=$(duration ${col2} 86400)}
didn't properly assign the returned value from the duration function, so function changed to print the value... to be assigned to he variable.